set category API response limit to 1
This fixes a problem when there are multiple categories with the same name.
The current implementation will then return the result as an array of categories,
whereas the code expects an object. Thus the category id can't be set.
This fix limits the response from the API to 1 and will therefor respond with an object.